Add auth on udaSB component
